+++ app/src/main/java/eu/decentnewsroom/bookshelf/ui/settings/SettingsUiState.kt
@@ -4,6 +4,7 @@ import eu.decentnewsroom.bookshelf.data.nostr.BookshelfSyncState
 import eu.decentnewsroom.bookshelf.data.nostr.NostrSignerSession
 import eu.decentnewsroom.bookshelf.data.nostr.RelayConfiguration
 import eu.decentnewsroom.bookshelf.data.reader.ReaderPreferences
+import eu.decentnewsroom.bookshelf.data.reader.OfflineBookCacheStats
 import eu.decentnewsroom.bookshelf.data.rendering.ChapterHtmlCacheStats
 import eu.decentnewsroom.bookshelf.data.ratings.BookRatingCacheStats
 
@@ -19,6 +20,7 @@ data class SettingsUiState(
     val pendingReviewCount: Int = 0,
     val chapterCacheStats: ChapterHtmlCacheStats = ChapterHtmlCacheStats(),
     val ratingCacheStats: BookRatingCacheStats = BookRatingCacheStats(),
+    val offlineBookCacheStats: OfflineBookCacheStats = OfflineBookCacheStats(),
     val isRefreshingStats: Boolean = false,
     val isRetrying: Boolean = false,
     val message: String? = null,